+++ LethalMissions/Networking/NetworkHandler.cs
@@ -37,7 +37,6 @@ namespace LethalMissions.Networking
         [ServerRpc(RequireOwnership = false)]
         public void SyncMissionsServerRpc(string serializedMissions)
         {
-            Debug.Log($"LethalMissions - SyncMissionsServerRpc: {serializedMissions}");
             currentSerializedMissions = serializedMissions;
             SyncMissionsClientRpc(serializedMissions);
         }
@@ -49,7 +48,7 @@ namespace LethalMissions.Networking
         [ClientRpc]
         public void SyncMissionsClientRpc(string serializedMissions)
         {
-            Debug.Log($"LethalMissions - SyncMissionsClientRpc: {serializedMissions}");
+
             currentSerializedMissions = serializedMissions;
             Plugin.MissionManager.SyncMissions(serializedMissions);
         }
@@ -61,7 +60,6 @@ namespace LethalMissions.Networking
         [ServerRpc(RequireOwnership = false)]
         public void RequestMissionsServerRpc(ulong clientId)
         {
-            Debug.Log($"LethalMissions - RequestMissionsServerRpc: {clientId}");
             if (string.IsNullOrEmpty(currentSerializedMissions))
             {
                 Debug.LogError("currentSerializedMissions is null or empty");
@@ -80,32 +78,9 @@ namespace LethalMissions.Networking
         {
             if (NetworkManager.Singleton.LocalClientId == targetClientId)
             {
-                Debug.Log($"LethalMissions - SendMissionsClientRpc: {serializedMissions}");
                 Plugin.MissionManager.SyncMissions(serializedMissions); // Pass the serialized missions to the MissionManager
             }
         }
 
-
-        [ServerRpc(RequireOwnership = false)]
-        public void SyncValvesServerRpc()
-        {
-            Debug.Log($"LethalMissions - SyncValvesServerRpc");
-            SyncValvesClientRpc();
-        }
-
-
-        [ClientRpc]
-        public void SyncValvesClientRpc()
-        {
-            Debug.Log($"LethalMissions - SyncValvesClientRpc");
-            SteamValveHazard[] allValves = FindObjectsOfType<SteamValveHazard>();
-            int N = allValves.Length == 1 ? 1 : new System.Random().Next(1, allValves.Length + 1);
-            for (int i = 0; i < N; i++)
-            {
-                allValves[i].valveCrackTime = 0.001f;
-                allValves[i].valveBurstTime = 0.01f;
-                allValves[i].triggerScript.interactable = true;
-            }
-        }
     }
 }